requirements.txt
argh==0.26.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
bleach==3.0.2
certifi==2018.10.15
cffi==1.11.5
channels==2.1.5
chardet==3.0.4
colorama==0.4.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
daphne==2.2.2
Django==2.1.2
django-cors-middleware==1.3.1
django-markdownify==0.8.0
django-pipeline==1.6.14
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
future==0.17.0
html2text==2018.1.9
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
injector==0.14.1
lxml==4.2.5
Markdown==3.0.1
parsel==1.5.1
patool==1.12
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-scrapyd-api==2.1.2
pytz==2018.6
queuelib==1.5.0
requests==2.20.0
Scrapy==1.5.1
scrapyd==1.2.0
service-identity==17.0.0
six==1.11.0
sqlparse==0.2.4
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24
urwid==2.0.1
uWSGI==2.0.17.1
w3lib==1.19.0
webencodings==0.5.1
zope.interface==4.6.0
